backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
PyMuPDF==1.23.8
docx2txt==0.8
python-dotenv==1.0.0
nltk==3.8.1
requests==2.31.0
Werkzeug==2.3.7
python-docx==0.8.11
groq==0.3.0